+++ data/shaders/piece/piece_no_normal.frag
@@ -2,14 +2,14 @@
 
 in vec2 v_texture_coordinate;
 in vec3 v_normal;
-in flat int v_entity_id;
+in flat float v_entity_id;
 in vec3 v_fragment_position;
 in vec4 v_fragment_position_light_space;
 in vec3 v_light_position;
 in vec3 v_view_position;
 
 layout(location = 0) out vec4 fragment_color;
-layout(location = 1) out int entity_id;
+layout(location = 1) out float entity_id;
 
 uniform sampler2D u_shadow_map;
